+++ CHANGELOG.md
@@ -5,8 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [3.1.0] - 2025-04-30
+### Added
+- support for B2B carts using the first company contact and location of a customer if set
+  (requires Shopify "new customer accounts" and `@shopgate/shopify-user` v2.3.0 or higher)
 
-## [3.0.4]
+### Fixed
+- line items that are out of stock are not left in the cart with 0 amount anymore (introduced in 3.0.4 with the API
+  upgrade)
+
+## [3.0.4] - 2025-04-02
 ### Changed
 - uses Shopify Storefront API 2025-01
 - uses Shopify REST Admin API 2025-01
@@ -148,7 +156,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - logging request duration and other info when requesting the Shopify API
 
-[Unreleased]: https://github.com/shopgate/ext-shopify-cart/compare/v3.0.4...HEAD
+[Unreleased]: https://github.com/shopgate/ext-shopify-cart/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/shopgate/ext-shopify-cart/compare/v3.0.4...v3.1.0
 [3.0.4]: https://github.com/shopgate/ext-shopify-cart/compare/v3.0.3...v3.0.4
 [3.0.3]: https://github.com/shopgate/ext-shopify-cart/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/shopgate/ext-shopify-cart/compare/v3.0.1...v3.0.2